Refactor : RacingCars 객체 생성을 한곳에서 관리 할 수 있도록 수정
